Point class was created
Static keyword was removed because functions were invisible for Main
The Project:
There is a Point_xy structure that hides the x and the y.
There is a Point structure that has this Point_xy structure and public functions.
	new, free, copy, toString, setters and getters
In the Main function there is a little test to check.